wip: split filter into multiple sub components
